requirements.txt
asgiref==3.2.10
Django==3.0.8
django-autocomplete-light==3.5.1
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
